requirements.txt
requests>=2.25.0
numpy>=1.20.0
scipy>=1.6.0
networkx>=2.5.0
plotly>=4.14.0
scikit-learn>=0.24.0
flask>=2.0.0
biopython>=1.78
gunicorn>=20.1.0